Add teleport helper that checks the configuration for blacklisted blocks.
This helper is the basis for all other Sponge teleport helper filters. Other plugins can use this blacklist by supplying it, with other filters, to the teleport helper. It caches block states and types, so a hook on reload has been added to invalidate this cache.

The configuration is available in the global config under "teleport-helper".